# backend/src/openapi.yaml
openapi: 3.1.0
info:
  title: Library Demo API
  version: 1.0.0

paths:
  /users:
    get:
      summary: List users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

  /users/{id}:
    get:
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found

  /media:
    get:
      summary: List media
      responses:
        '200':
          description: A list of media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'

    post:
      summary: Create media
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'

  /media/{id}:
    get:
      summary: Get media by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A media item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'

    patch:
      summary: Update media
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'

    delete:
      summary: Delete media
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted

  /media-status:
    get:
      summary: List media status records
      responses:
        '200':
          description: A list of media status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaStatus'

    post:
      summary: Create media status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaStatusInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaStatus'

  /media-status/{id}:
    get:
      summary: Get media status by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A media status record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaStatus'

    patch:
      summary: Update media status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaStatusUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaStatus'

  /media/read-all:
    post:
      summary: Read all media (bulk query)
      description: Returns all media, optionally filtered
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                active:
                  type: boolean
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: List of media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'

  /media-status/read-all:
    post:
      summary: Read all media status records (bulk query)
      description: Returns all media status records, optionally filtered
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mediaIds:
                  type: array
                  items:
                    type: integer
                userIds:
                  type: array
                  items:
                    type: integer
                type:
                  type: string
                  enum: [ready, loaned, returned]
                fromDate:
                  type: string
                  format: date-time
                toDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: List of media status records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaStatus'

components:
  schemas:

    Locale:
      type: string
      enum: [en, es, fr]

    User:
      type: object
      required:
        - id
        - active
        - login
        - password
        - type
      properties:
        id:
          type: integer
        active:
          type: boolean
        login:
          type: string
        password:
          type: string
        type:
          type: string
          enum: [admin, patron]
        locale:
          $ref: '#/components/schemas/Locale'
        phone:
          type: string

    Media:
      type: object
      required:
        - id
        - type
        - title
        - creator
        - added
        - active
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [book]
        title:
          type: string
        creator:
          type: string
        desc:
          type: string
        tags:
          type: array
          items:
            type: string
        added:
          type: string
          format: date-time
        active:
          type: boolean
        cost:
          type: number

    MediaInput:
      type: object
      required:
        - type
        - title
        - creator
        - added
        - active
      properties:
        type:
          type: string
          enum: [book]
        title:
          type: string
        creator:
          type: string
        desc:
          type: string
        tags:
          type: array
          items:
            type: string
        added:
          type: string
          format: date-time
        active:
          type: boolean
        cost:
          type: number

    MediaUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [book]
        title:
          type: string
        creator:
          type: string
        desc:
          type: string
        tags:
          type: array
          items:
            type: string
        added:
          type: string
          format: date-time
        active:
          type: boolean
        cost:
          type: number

    MediaStatus:
      type: object
      required:
        - id
        - mediaId
        - type
        - user
        - date
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        type:
          type: string
          enum: [ready, loaned, returned]
        user:
          type: integer
        date:
          type: string
          format: date-time

    MediaStatusInput:
      type: object
      required:
        - mediaId
        - type
        - user
        - date
      properties:
        mediaId:
          type: integer
        type:
          type: string
          enum: [ready, loaned, returned]
        user:
          type: integer
        date:
          type: string
          format: date-time

    MediaStatusUpdate:
      type: object
      properties:
        mediaId:
          type: integer
        type:
          type: string
          enum: [ready, loaned, returned]
        user:
          type: integer
        date:
          type: string
          format: date-time